Use unique namespace for each test
